Move color utils to colors.stories folder
