+++ storybook/stand-alone-documentation/colors/colors-description/colors-description.jsx
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
-import { Frame, ColorDescription } from "vibe-storybook-components";
-import { colorsHashMap } from "../../../../utils/colors-vars-map";
+import { ColorDescription } from "vibe-storybook-components";
+import { colorsHashMap } from "../utils/colors-vars-map";
 
 const colorsWithBorder = new Set([
   "text-color-on-inverted",
