+++ storybook/stand-alone-documentation/colors/utils/colors-vars-map.ts
